app/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlalchemy==2.0.23
alembic==1.11.1
pydantic==1.10.7
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pandas==2.2.2
joblib==1.4.2
scikit-learn==1.5.2
tensorflow==2.16.1
python-dotenv==1.0.0
email-validator
